Match warpstonelift_update via flat-sum K-grouping
Drop the inner parens on the squad-member pointer load so off+0x100 folds
onto the index (addi off,256; lwzx) matching target's fold-onto-index isel
(#112). 98.92 -> 100%.

(cherry picked from commit 598666a09ffb4e28ae37433c07cf4cf5914b3111)

+++ src/main/dll/SH/dll_01AF_shswaplift.c
@@ -74,7 +74,7 @@ void warpstonelift_update(u8* obj)
         off = 0;
         for (i = 0; i < count; i++)
         {
-            char* o = *(char**)((int)p + (off + 0x100));
+            char* o = *(char**)(p + off + 0x100);
             if (((GameObject*)o)->anim.classId == 1)
             {
                 found = 1;
